+++ keytool/src/main/scala/org/constellation/keytool/KeyUtils.scala
@@ -56,6 +56,7 @@ object KeyUtils extends StrictLogging {
   private val PrivateKeyHexPrefix: String =
     "30818d020100301006072a8648ce3d020106052b8104000a047630740201010420"
   private val PrivateKeyHexPrefixLength: Int = PrivateKeyHexPrefix.length
+  private val secp256kHexIdentifier: String = "a00706052b8104000aa144034200"
 
   /**
     * Simple Bitcoin like wallet grabbed from some stackoverflow post
@@ -200,9 +201,19 @@ object KeyUtils extends StrictLogging {
   def hexToPublicKey(hex: String): PublicKey =
     bytesToPublicKey(hex2bytes(PublicKeyHexPrefix + hex))
 
-  def privateKeyToHex(privateKey: PrivateKey): String = {
+  /**
+    * Full hex has following format:
+    * PRIVATEKEYHEXa00706052b8104000aa144034200PUBLICKEYHEX where a00706052b8104000aa144034200 stands for secp256k identifier
+    */
+  def privateKeyToFullHex(privateKey: PrivateKey): String = {
     val hex = bytes2hex(privateKey.getEncoded)
     hex.slice(PrivateKeyHexPrefixLength, hex.length)
+  }
+
+  def privateKeyToHex(privateKey: PrivateKey): String = {
+    val fullHex = privateKeyToFullHex(privateKey)
+    val privateHex = fullHex.split(secp256kHexIdentifier).head
+    privateHex
   }
 
   def hexToPrivateKey(hex: String): PrivateKey =