+++ src/test/scala/org/constellation/util/wallet/KeyToolTest.scala
@@ -7,9 +7,9 @@ import cats.effect.IO
 import org.constellation.Fixtures
 import org.constellation.Fixtures._
 import org.constellation.keytool.{KeyStoreUtils, KeyTool, KeyUtils}
-import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach}
 import org.scalatest.flatspec.AsyncFlatSpecLike
 import org.scalatest.matchers.should.Matchers
+import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach}
 
 class KeyToolTest extends AsyncFlatSpecLike with Matchers with BeforeAndAfterAll with BeforeAndAfterEach {
   val keyTool = KeyTool
@@ -21,6 +21,11 @@ class KeyToolTest extends AsyncFlatSpecLike with Matchers with BeforeAndAfterAll
     s"--storepass=${Fixtures.storepass}",
     s"--keypass=${Fixtures.keypass}"
   )
+
+  override protected def beforeEach(): Unit = {
+    File(savedKeystorePath).delete(true)
+    super.beforeEach()
+  }
 
   "KeyStoreUtils" should "load keypair successfully" ignore {
     val loadKp =
@@ -64,5 +69,13 @@ class KeyToolTest extends AsyncFlatSpecLike with Matchers with BeforeAndAfterAll
 
   "keyPairFromPemStr" should "generate keypair from PEM strings" in {
     assert(KeyUtils.keyPairFromPemStr(Fixtures.privateKeyStr, Fixtures.pubKeyStr).isInstanceOf[KeyPair])
+  }
+
+  "KeyTool" should "convert private key to hex and back from hex" in {
+    val keypair = KeyUtils.makeKeyPair()
+    val privateKey = keypair.getPrivate
+    val hex = KeyUtils.privateKeyToFullHex(privateKey)
+    val fromHex: PrivateKey = KeyUtils.hexToPrivateKey(hex)
+    assert(fromHex.equals(privateKey))
   }
 }