+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Possibility to migrate existing KeyStore to use storepass for both KeyStore and KeyPair
+- Possibility to export private key from KeyStore in hexadecimal format
 
 ## [v2.16.2] 2020-10-09
 ### Changed